% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk.R
\name{check_generic}
\alias{check_generic}
\alias{chk_dummy}
\alias{chk_blank}
\alias{chk_equals}
\alias{chk_range}
\alias{chk_pattern}
\alias{chk_length}
\alias{chk_miss}
\alias{chk_nmiss}
\alias{chk_unique}
\alias{chk_date_yyyymmdd}
\alias{chk_date_yyyymm}
\alias{chk_date_yyyy}
\alias{chk_ascii}
\alias{chk_values}
\title{Generic Checking Functions}
\usage{
chk_dummy(x)

chk_blank(x)

chk_equals(x, val)

chk_range(x, min, max)

chk_pattern(x, pattern)

chk_length(x, len)

chk_miss(x, miss = getOption("testdat.miss_text"))

chk_nmiss(x, miss = getOption("testdat.miss_text"))

chk_unique(x)

chk_date_yyyymmdd(x)

chk_date_yyyymm(x)

chk_date_yyyy(x)

chk_ascii(x)

chk_values(x, ..., miss = TRUE)
}
\arguments{
\item{x}{vector to check}

\item{val}{value for equality check}

\item{min}{minimum value for range check}

\item{max}{maximum value for range check}

\item{pattern}{pattern to look for as defined in
\link[stringr:str_detect]{str_detect()}}

\item{len}{maximum string length for checking string variables}

\item{miss}{vector of values to be treated as missing}

\item{...}{vectors of valid values}
}
\value{
A logical vector flagging records that have passed or failed the check
}
\description{
These functions provide common, simple data checks.
}
\seealso{
\link[=check_helper]{Checking Helper Functions}
}
