% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk.R
\name{check_helper}
\alias{check_helper}
\alias{chk_filter}
\alias{chk_filter_vars}
\alias{chk_filter_all}
\alias{chk_filter_any}
\title{Checking Helper Functions}
\usage{
chk_filter(.dat, .var, .func, .flt = TRUE, .args = list())

chk_filter_vars(.dat, .vars, .func, .flt = TRUE, .args = list())

chk_filter_all(.dat, .vars, .func, .flt = TRUE, .args = list())

chk_filter_any(.dat, .vars, .func, .flt = TRUE, .args = list())
}
\arguments{
\item{.dat}{a \code{tbl}}

\item{.func}{a function to use use for checking. The provided function should take a single vector as input and return a logical vector.}

\item{.flt}{a logical expression to be applied as a filter for the check. This will be evaluated in the context of .dat}

\item{.args}{a list of additional arguments to be added to the function calls.}

\item{.vars}{a list of columns to check generated by \link[dplyr:vars]{vars()}}
}
\value{
A logical vector flagging records that have passed or failed the check, with \code{NA} where records do not meet the filter
}
\description{
These helper functions allowing easy checking over multiple columns and filtering to subsets of the data.
}
