% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprec-chk.R
\name{chk-defunct}
\alias{chk-defunct}
\alias{chk_filter_vars}
\alias{chk_filter_where}
\title{Defunct checking functions}
\usage{
chk_filter_vars(data, vars, func, flt = TRUE, args = list())

chk_filter_where(data, where, func, flt = TRUE, args = list())
}
\arguments{
\item{where}{<\code{\link[=dplyr_tidy_select]{tidy-select}}> Columns to check.}
}
\value{
A logical vector flagging records that have passed or failed the
check.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#defunct}{\figure{lifecycle-defunct.svg}{options: alt='[Defunct]'}}}{\strong{[Defunct]}}

These functions are defunct.
\itemize{
\item \code{chk_filter_where()} works exactly like \code{chk_filter_all()}. When testdat
used \code{dplyr::vars()} as standard \code{chk_filter_where()} provided an alternative
interface using \code{\link[=dplyr_tidy_select]{tidy-select}}.
}
}
\seealso{
\verb{chk_*()} functions such as \code{\link[=chk-values]{chk_values()}}

\link[=chk-helper]{Checking Helper Functions}
}
\keyword{internal}
