% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprec-chk.R
\name{chk-deprec}
\alias{chk-deprec}
\alias{chk_length}
\alias{chk_miss}
\alias{chk_nmiss}
\title{Deprecated checking functions}
\usage{
chk_length(x, len)

chk_miss(x, miss = getOption("testdat.miss_text"))

chk_nmiss(x, miss = getOption("testdat.miss_text"))
}
\arguments{
\item{x}{vector to check}

\item{len}{maximum string length for checking string variables}

\item{miss}{vector of values to be treated as missing}
}
\value{
A logical vector flagging records that have passed or failed the check
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

These functions are deprecated. See \link[=chk-generic]{Generic Checking Functions} for current checking functions.
}
\seealso{
\link[=chk-generic]{Generic Checking Functions}
}
\keyword{internal}
