% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk.R
\name{chk-dummy}
\alias{chk-dummy}
\alias{chk_dummy}
\title{Checks: dummy}
\usage{
chk_dummy(x)
}
\arguments{
\item{x}{A vector to check.}
}
\value{
A logical vector flagging records that have passed or failed the
check.
}
\description{
These functions provide common, simple data checks.
}
\examples{

chk_dummy(LETTERS)

}
\seealso{
\link[=chk-helper]{Checks: data frame helpers}

Other vector checks: 
\code{\link{chk-dates}},
\code{\link{chk-labels}},
\code{\link{chk-patterns}},
\code{\link{chk-text}},
\code{\link{chk-uniqueness}},
\code{\link{chk-values}}
}
\concept{vector checks}
