% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expect-make.R
\name{chk-expect}
\alias{chk-expect}
\alias{expect_date_yyyy}
\alias{expect_date_yyyymm}
\alias{expect_date_yyyymmdd}
\alias{expect_max_length}
\alias{expect_text_miss}
\alias{expect_text_nmiss}
\title{Expectations: auto-generated}
\usage{
expect_date_yyyy(var, flt = TRUE, data = get_testdata())

expect_date_yyyymm(var, flt = TRUE, data = get_testdata())

expect_date_yyyymmdd(var, flt = TRUE, data = get_testdata())

expect_max_length(var, len, flt = TRUE, data = get_testdata())

expect_text_miss(
  var,
  miss = getOption("testdat.miss_text"),
  flt = TRUE,
  data = get_testdata()
)

expect_text_nmiss(
  var,
  miss = getOption("testdat.miss_text"),
  flt = TRUE,
  data = get_testdata()
)
}
\arguments{
\item{var}{an unquoted variable name}

\item{flt}{a filter specifying a subset of the dataset to test}

\item{data}{data frame to test. The \link[=global-data]{global test data} is used
by default.}

\item{len}{maximum string length for checking string variables}

\item{miss}{vector of values to be treated as missing}
}
\description{
These expectations are auto-generated from the \verb{chk_*()} functions of the
same name.
}
\examples{
sales <- data.frame(
  sale_id = 1:5,
  date = c("20200101", "20200101", "20200102", "20200103", "2020003"),
  sale_price = c(10, 20, 30, 40, -1),
  book_title = c(
    "Phenomenology of Spirit",
    NA,
    "Critique of Practical Reason",
    "Spirit of Trust",
    "Empiricism and the Philosophy of Mind"
  ),
  stringsAsFactors = FALSE
)

# Validate sales
\dontrun{
expect_text_nmiss(book_title, data = sales) # Titles non-missing
expect_date_yyyymmdd(date, data = sales) # Dates in correct format
}
}
\seealso{
\link[=chk-generic]{Generic Checking Functions}
}
