% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expect-data.R
\name{conditional-expectations}
\alias{conditional-expectations}
\alias{expect_base}
\alias{expect_cond}
\title{Expectations: consistency checks}
\usage{
expect_base(var, base, miss = getOption("testdat.miss"),
  data = get_testdata())

expect_cond(cond1, cond2, data = get_testdata())
}
\arguments{
\item{var}{an unquoted variable name}

\item{base}{condition for missing check}

\item{miss}{a vector of values to be treated as missing. The
\link[=testdat]{testdat.miss} option is used by default}

\item{data}{data frame to test. The \link[=global-data]{global test data} is used
by default.}

\item{cond1}{first condition for consistency check}

\item{cond2}{second condition for consistency check}
}
\description{
These functions test whether multiple conditions coexist.
}
\details{
\code{expect_cond} checks the coexistence of two conditions. It can be read as "if
\code{cond1} then \code{cond2}".

\code{expect_base} is a special case that checks missing data against a specified
condition. It can be read as "if \code{base} then \code{var} not missing, if not \code{base}
then \code{var} missing".
}
\seealso{
Other data expectations: \code{\link{datacomp-expectations}},
  \code{\link{generic-expectations}},
  \code{\link{value-expectations}}
}
