% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporter-zzz.R
\name{data-accessors}
\alias{data-accessors}
\alias{set_testdata}
\alias{get_testdata}
\title{Get/set test data}
\usage{
set_testdata(data)

get_testdata()
}
\arguments{
\item{data}{data to be used}
}
\description{
Changes global test data to that specified. Used by data expectation functions
}
\keyword{internal}
