% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expect-data.R
\name{data-params}
\alias{data-params}
\title{Expectation params}
\arguments{
\item{var}{an unquoted variable name}

\item{vars}{a list of columns generated by \link[dplyr:vars]{vars()}}

\item{flt}{a filter specifying a subset of the dataset to test}

\item{miss}{a vector of values to be treated as missing. The
\link[=testdat]{testdat.miss} option is used by default}

\item{data}{data frame to test. The \link[=global-data]{global test data} is used
by default.}
}
\description{
Expectation params
}
\keyword{internal}
