\name{detect_outliers}
\alias{detect_outliers}
\title{Functions to import: scores from outliers package}
\usage{
detect_outliers(dat, threshold = 1.5, ...)
}
\arguments{
  \item{dat}{data.frame input}

  \item{threshold}{Threshold value}

  \item{...}{Further args}
}
\value{
Returns an index with outliers to investigate
}
\description{
Functions to import: scores from outliers package
}

