% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporter-excel.R
\docType{data}
\name{ExcelReporter}
\alias{ExcelReporter}
\title{Excel reporter: gather all test results along with elapsed time and
file information.}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ExcelReporter
}
\description{
This reporter gathers all results, adding additional information such as
test elapsed time, and test filename if available. Very useful for reporting.
}
\keyword{datasets}
