% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expect-exclusive.R
\name{exclusivity-expectations}
\alias{exclusivity-expectations}
\alias{expect_exclusive}
\title{Expectations: exclusivity checks}
\usage{
expect_exclusive(
  vars,
  exc_vars,
  exc_val = 1,
  flt = TRUE,
  data = get_testdata()
)
}
\arguments{
\item{vars}{a list of columns generated by \link[dplyr:vars]{vars()}}

\item{exc_vars}{The variable set to check, specified using \code{\link[=vars]{vars()}}. This
should include all variables specified in the \code{vars} argument}

\item{exc_val}{The value to check for exclusivity (default: \code{1})}

\item{flt}{a filter specifying a subset of the dataset to test}

\item{data}{data frame to test. The \link[=global-data]{global test data} is used
by default.}
}
\description{
\code{expect_exclusive} tests one set of variables (\code{exc_vars}) against another
(\code{vars}) with reference to a value (\code{exc_val}) and fails if any observation
has \emph{all} of the variables in \code{vars} equal to \code{exc_val} and and \emph{any} of the
variables in \code{exc_vars} (aside from those which are also in \code{vars}) equal to
\code{exc_val}.
}
\examples{

my_q_block <- data.frame(
  resp_id = 1:5, # Unique to respondent
  q10_1 = c(1, 1, 0, 0, 0),
  q10_2 = c(0, 1, 0, 0, 0),
  q10_3 = c(0, 0, 1, 0, 0),
  q10_98 = c(1, 0, 0, 1, 0), # None of the above
  q10_99 = c(0, 0, 0, 0, 1)  # Item not answered
)

# Make sure that if "None of the above" and "Item skipped" are selected
# none of the other question options are selected:
\dontrun{expect_exclusive(
  vars(q10_98, q10_99),
  vars(starts_with("q10_")),
  data = my_q_block
)}
}
\seealso{
Other data expectations: 
\code{\link{chk-expectations}},
\code{\link{conditional-expectations}},
\code{\link{datacomp-expectations}},
\code{\link{generic-expectations}},
\code{\link{uniqueness-expectations}}
}
\concept{data expectations}
