% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expect-deprec.R
\name{expect-deprec}
\alias{expect-deprec}
\alias{expect_func}
\title{Deprecated expectation functions}
\usage{
expect_func(var, ...)
}
\arguments{
\item{var}{an unquoted variable name}

\item{...}{arguments to pass to \code{expect_allany()}}
}
\description{
\Sexpr[results=rd, stage=render]{testdat:::lifecycle("soft-deprecated")}

These functions are deprecated.

See \link[=generic-expectations]{Generic Expectation Functions} for current expectation functions.
}
\seealso{
\link[=generic-expectations]{Generic Expectation Functions}
}
