% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expect-data.R
\name{expect_custom}
\alias{expect_custom}
\title{Extension of 'expect' to allow inclusion of custom fields}
\usage{
expect_custom(ok, failure_message, info = NULL, srcref = NULL, ...)
}
\arguments{
\item{ok}{Was the expectation successful?}

\item{failure_message}{What message should be shown if the expectation was
not successful?}

\item{info}{Additional information. Included for backward compatibility only
and new expectations should not use it.}

\item{srcref}{Only needed in very rare circumstances where you need to
forward a srcref captured elsewhere.}

\item{...}{Additional data to be added to a list in the \code{custom}
attribute of the resulting expectation.}
}
\description{
Use expect_custom to allow inclusion of arbitrary data in expectation
results. Additional data is stored in a list in an attribute called
\code{custom} in the resulting expectation. THis allows data expectations to
store information about the number of failed and successful cases for
reporting of test results.
}
\examples{
x <- expect_custom(TRUE, "Test", extra_data = 1:5, more_data = "Hello")
str(x)
}
