% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expect-make.R
\name{expect_make}
\alias{expect_make}
\title{Make an expectation from a check function}
\usage{
expect_make(func, func_desc = NULL, vars = FALSE, all = TRUE,
  env = caller_env())
}
\arguments{
\item{func}{A function whose first argument takes a vector to check, and
returns a logical vector of the same length with the results.}

\item{func_desc}{A character function description to use in th}

\item{vars}{Should the generated function take a single variable name or a \code{vars()} object?}

\item{all}{Function to use to combine results for each vector.}

\item{env}{The parent environment of the function, defaults to the calling
environment of \code{expect_make()}}
}
\value{
An \code{expect_*()} style function.
}
\description{
\code{expect_make()} creates an expection from a vectorised checking function to
allow simple generation of domain specific data checks.
}
\examples{
expect_make(chk_pattern)
expect_make(chk_values)
}
