% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expectation.R
\name{filter_expect}
\alias{filter_expect}
\title{Filter data to expectation result}
\usage{
filter_expect(data, expect_function, ..., not = TRUE)
}
\description{
\Sexpr[results=rd, stage=render]{testdat:::lifecycle("experimental")}
}
