% Generated by roxygen2 (4.0.1): do not edit by hand
\name{fix_date_continuity}
\alias{fix_date_continuity}
\title{Add NA rows to datestream datasets}
\usage{
fix_date_continuity(dat, index, form = ymd, granularity = "days")
}
\arguments{
\item{dat}{data frame with a date column that should contain a continuous stream of dates}

\item{index}{index of column in \code{dat} containing dates}

\item{form}{a lubridate parsing function in the \code{ymd} family, specifying the format the dates in \code{dat} are expected to be in. Default \code{ymd}.}

\item{granularity}{what are the datestream units? Choose one of \code{'secs'}, \code{'mins'}, \code{'hours'}, \code{'days'}, \code{'weeks'}, \code{'years'}. Default \code{'days'}.}
}
\value{
a data frame, sorted by the dates in \code{dat[,index]}, with the missing observations added. Dates in \code{dat[,index]} will be formatted as \code{POSIXct}, due to conversion with \code{lubridate}.
}
\description{
For a dataset that should contain at least one observation for each day
(or second, minute, hour, etc), add NA rows for the missing observations
}

