% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expect-data.R
\name{generic-expectations}
\alias{generic-expectations}
\alias{expect_all}
\alias{expect_any}
\alias{expect_func}
\title{Expectations: generic expectations}
\usage{
expect_all(...)

expect_any(...)

expect_func(var, ...)
}
\arguments{
\item{var}{an unquoted variable name}

\item{func}{a function that takes a vector as the first argument and returns
a logical vector of the same length showing whether an element passed or
failed}

\item{args}{a named list of arguments to pass to \code{func}}
}
\description{
These functions allow for testing of a dataset using an arbitrary function.
}
\seealso{
\link{check_generics} for a set of generic checking functions

Other data expectations: \code{\link{conditional-expectations}},
  \code{\link{datacomp-expectations}},
  \code{\link{value-expectations}}
}
