% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporter-zzz.R
\name{global-data}
\alias{global-data}
\alias{set_testdata}
\alias{get_testdata}
\alias{with_testdata}
\title{Get/set test data}
\usage{
set_testdata(data, quosure = TRUE)

get_testdata()

with_testdata(data, code, quosure = TRUE)
}
\arguments{
\item{data}{Data to be used.}

\item{quosure}{If \code{TRUE}, the default, the dataset is stored as a quosure and
\code{get_testdata()} will return the current state of the dataset.

If \code{FALSE}, the dataset will be copied and \code{get_testdata()} will return the
state of the dataset at the time \code{set_testdata()} was called.}

\item{code}{Code to execute.}
}
\description{
A global test data set is used to avoid having to re-specify the testing
dataset in every test. These functions get and set the global data or set the
data for the current context.
}
\examples{
set_testdata(mtcars)
head(get_testdata())

with_testdata(iris, {
  x <- get_testdata()
  print(head(x))
})
}
