% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spss_to_r.R
\name{spss_to_r}
\alias{spss_to_r}
\title{modified from 
https://github.com/lebebr01/SPSStoR/blob/master/R/selectif_to_r.r}
\usage{
spss_to_r(x)
}
\description{
modified from 
https://github.com/lebebr01/SPSStoR/blob/master/R/selectif_to_r.r
}
