% Generated by roxygen2 (4.0.1): do not edit by hand
\name{test_dupname}
\alias{test_dupname}
\title{Test for duplicate column names}
\usage{
test_dupname(dat)
}
\arguments{
\item{dat}{input dataset. Currently only supports \code{data.frame}.}
}
\description{
This test will check to see that there are no duplicate column names in a data frame.
Will return \code{TRUE} if there is a duplicate column name, otherwise \code{FALSE}.
}

