% Generated by roxygen2 (4.0.1): do not edit by hand
\name{test_long}
\alias{test_long}
\title{Valid longitude
Test for valid longitude. Between -180 and 180}
\usage{
test_long(x)
}
\arguments{
\item{x}{A longitude}
}
\description{
Valid longitude
Test for valid longitude. Between -180 and 180
}
\examples{
\dontrun{
expect_true(test_lat(dat$longitude))
}
}

