% Generated by roxygen2 (4.0.1): do not edit by hand
\name{test_white_spaces}
\alias{test_white_spaces}
\title{test_for_white_spaces}
\usage{
test_white_spaces(text)
}
\arguments{
\item{text}{Input text. Usually this would be a \code{character} field.}
}
\description{
Tests white spaces that could end up padding text
}
\examples{
\dontrun{
test_white_spaces(" Karthik Ram")
test_white_spaces(" Karthik Ram ")
}
}

