% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testdat.R
\docType{package}
\name{testdat-package}
\alias{testdat}
\alias{testdat-package}
\title{R package for data unit testing}
\description{
An adaptation of \link{testthat} for data. See vigenttes for details.
}
\section{Options}{

\itemize{
\item \code{testdat.miss}: A vector of values to consider missing (default: \code{c(NA, "")}).
\item \code{testdat.miss_text}: A vector of values to consider missing in text variables
(default: \code{c("error", "null", "0", ".", "-", ",", "na", "#n/a", "", NA)}).
\item \code{testdat.stop_on_fail}: Should an expectation raise an error on failure?
Useful for interactive use of expectation functions (default: \code{TRUE}).
}
}

\author{
\strong{Maintainer}: Danny Smith \email{daniel.smith@srcentre.com.au}

Other contributors:
\itemize{
  \item The Social Research Centre [copyright holder]
}

}
\keyword{internal}
