% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporter-data.R
\name{testdat_results}
\alias{testdat_results}
\title{Create a `testdat_results` object from the test results
as stored in the DataReporter results field.}
\usage{
testdat_results(results)
}
\arguments{
\item{results}{a list as stored in DataReporter}
}
\value{
its list argument as a `testdat_results` object
}
\description{
Create a `testdat_results` object from the test results
as stored in the DataReporter results field.
}
\seealso{
DataReporter
}
\keyword{internal}
