% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expect-unique.R
\name{uniqueness-expectations}
\alias{uniqueness-expectations}
\alias{expect_unique}
\alias{expect_unique_across}
\alias{expect_unique_combine}
\title{Expectations: uniqueness checks}
\usage{
expect_unique(vars, flt = TRUE, data = get_testdata())

expect_unique_across(vars, flt = TRUE, data = get_testdata())

expect_unique_combine(vars, flt = TRUE, data = get_testdata())
}
\arguments{
\item{vars}{a list of columns generated by \link[dplyr:vars]{vars()}}

\item{flt}{a filter specifying a subset of the dataset to test}

\item{data}{data frame to test. The \link[=global-data]{global test data} is used
by default.}
}
\description{
These functions test variables for uniqueness:
\itemize{
\item \code{expect_unique} tests a set of variables (\code{vars}) and fails if the
variables collectively do not uniquely identify the observations
\item \code{expect_unique_across} tests a set of variables (\code{vars}) and fails if
each observation does not have different values in each of the variables
in \code{vars}.
\item \code{expect_unique_combine} tests a set of variables (\code{vars}) and fails if
any values appear more than once across all of them.
}
}
\examples{

student_fruit_preferences <- data.frame(
  student_id = 1:5,
  apple = c(1, 1, 1, 1, 1),
  orange = c(2, 3, 2, 3, 2),
  banana = c(3, 2, 3, 2, 3),
  phone1 = c(123, 456, 789, 987, 654),
  phone2 = c(345, 678, 987, 567, 000)
)

#' # Check that key is unique
expect_unique(vars(student_id), data = student_fruit_preferences)

# Check that every observation has a different value across a set of variables
expect_unique_across(vars(apple, orange, banana), data = student_fruit_preferences)

# Check that each phone number appears at most once
\dontrun{expect_unique_combine(vars(phone1, phone2), data = student_fruit_preferences)}
}
\seealso{
Other data expectations: 
\code{\link{conditional-expectations}},
\code{\link{datacomp-expectations}},
\code{\link{exclusivity-expectations}},
\code{\link{generic-expectations}},
\code{\link{value-expectations}}
}
\concept{data expectations}
