% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project-testdat.R
\name{use_testdat}
\alias{use_testdat}
\title{Setup project for data tests}
\usage{
use_testdat(path = rstudioapi::getActiveProject(), excelres = FALSE,
  exceldef = FALSE, build = FALSE)
}
\arguments{
\item{path}{the project path to use. Defaults to the current open project in
RStudio.}

\item{excelres}{set up to output results to Excel file}

\item{exceldef}{set up to read results from definitions file}

\item{build}{set up for build automation in RStudio. Different OS require
different build scripts - this should be either "win" or "linux" as
applicable}
}
\description{
\code{use_testdat()} sets up data testing infrastructure.
}
\details{
Creates a "test" sub-directory with an example test script \code{test-example.R}
and a script \code{_run-all.R} to run all associated tests. Optionally:
\itemize{
\item add files \code{def-tests.R} and \code{definitions.xlsx} to allow Excel-defined
tests. To use, add \code{source("def-tests.R")} to your main testing file.
\item set up to output results to an Excel file \code{results.xlsx} using the
\code{output_results_excel()} function.
\item add project build settings in R. This creates a batch file and modifies the
RStudio project so your tests can be run directly from the "Build" menu.
}
}
