% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expect-value.R
\name{value-expectations}
\alias{value-expectations}
\alias{expect_values}
\alias{expect_regex}
\alias{expect_range}
\title{Expectations: value checks}
\usage{
expect_values(
  var,
  ...,
  miss = getOption("testdat.miss"),
  data = get_testdata()
)

expect_regex(var, pattern, flt = TRUE, data = get_testdata())

expect_range(var, min, max, ..., flt = TRUE, data = get_testdata())
}
\arguments{
\item{var}{an unquoted variable name}

\item{...}{vectors of valid values}

\item{miss}{a vector of values to be treated as missing. The
\link[=testdat]{testdat.miss} option is used by default}

\item{data}{data frame to test. The \link[=global-data]{global test data} is used
by default.}

\item{pattern}{a regex to check}

\item{flt}{a filter specifying a subset of the dataset to test}

\item{min}{minimum value for range check}

\item{max}{maximum value for range check}
}
\description{
These functions test variable values:
\itemize{
\item \code{expect_values} tests one variable (\code{var}) against a set of values
(\code{...}) and fails if there are any other values in the variable
\item \code{expect_regex} tests one variable (\code{var}) against a regex pattern
(\code{pattern}) and fails if there are any values not conforming to that
pattern in the variable
\item \code{expect_range} tests one variable (\code{var}) against a range (defined by
\code{min}/\code{max} values) and fails if there are any values in the variable
outside this range
}
}
\examples{

# Check that all values in a variable fall into a given set
expect_values(vs, 0:1, data = mtcars)

# Check that all values in a variable conform to a certain regex pattern
expect_regex(vs, "[01]", data = mtcars)

# Check that all values in a variable fall into a given range
expect_range(mpg, 10, 40, data = mtcars)
}
\seealso{
Other data expectations: 
\code{\link{conditional-expectations}},
\code{\link{datacomp-expectations}},
\code{\link{exclusivity-expectations}},
\code{\link{generic-expectations}},
\code{\link{uniqueness-expectations}}
}
\concept{data expectations}
