% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk-deprec.R
\name{chk-deprec}
\alias{chk-deprec}
\alias{chk_length}
\alias{chk_miss}
\alias{chk_nmiss}
\title{Deprecated checking functions}
\usage{
chk_length(x, len)

chk_miss(x, miss = getOption("testdat.miss_text"))

chk_nmiss(x, miss = getOption("testdat.miss_text"))
}
\arguments{
\item{x}{vector to check}

\item{len}{maximum string length for checking string variables}

\item{miss}{vector of values to be treated as missing}
}
\value{
A logical vector flagging records that have passed or failed the check
}
\description{
\Sexpr[results=rd, stage=render]{testdat:::lifecycle("soft-deprecated")}

These functions are deprecated. See \link[=chk-generic]{Generic Checking Functions} for current checking functions.
}
\seealso{
\link[=chk-generic]{Generic Checking Functions}
}
