% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprec-chk.R
\name{chk-deprec}
\alias{chk-deprec}
\alias{chk_length}
\alias{chk_miss}
\alias{chk_nmiss}
\title{Deprecated checking functions}
\usage{
chk_length(x, len)

chk_miss(x, miss = getOption("testdat.miss_text"))

chk_nmiss(x, miss = getOption("testdat.miss_text"))
}
\arguments{
\item{x}{A vector to check.}

\item{len}{Maximum string length for checking string variables.}
}
\value{
A logical vector flagging records that have passed or failed the
check.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

These functions are deprecated.
}
\seealso{
\verb{chk_*()} functions such as \code{\link[=chk-values]{chk_values()}}

\link[=chk-helper]{Checking Helper Functions}
}
\keyword{internal}
