% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expect-make.R
\name{chk_expect}
\alias{chk_expect}
\alias{expect_date_yyyy}
\alias{expect_date_yyyymm}
\alias{expect_date_yyyymmdd}
\alias{expect_max_length}
\title{Generic Checking Expectations}
\usage{
expect_date_yyyy(var, flt = TRUE, data = get_testdata())

expect_date_yyyymm(var, flt = TRUE, data = get_testdata())

expect_date_yyyymmdd(var, flt = TRUE, data = get_testdata())

expect_max_length(var, len, flt = TRUE, data = get_testdata())
}
\description{
These expectations are auto-generated from the \code{chk_*()} functions of the
same name.
}
