% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expect-conditional.R
\name{conditional-expectations}
\alias{conditional-expectations}
\alias{expect_base}
\alias{expect_cond}
\title{Expectations: consistency checks}
\usage{
expect_base(
  var,
  base,
  miss = getOption("testdat.miss"),
  missing_valid = FALSE,
  data = get_testdata()
)

expect_cond(cond1, cond2, data = get_testdata())
}
\arguments{
\item{var}{an unquoted variable name}

\item{base}{condition for missing check}

\item{miss}{a vector of values to be treated as missing. The
\link[=testdat]{testdat.miss} option is used by default}

\item{missing_valid}{allow missing values for records meeting the condition.
This allows 'one way' base checks. This is \code{FALSE} by default.}

\item{data}{data frame to test. The \link[=global-data]{global test data} is used
by default.}

\item{cond1}{first condition for consistency check}

\item{cond2}{second condition for consistency check}
}
\description{
These functions test whether multiple conditions coexist.
}
\details{
\code{expect_cond} checks the coexistence of two conditions. It can be read as "if
\code{cond1} then \code{cond2}".

\code{expect_base} is a special case that checks missing data against a specified
condition. It can be read as "if \code{base} then \code{var} not missing, if not \code{base}
then \code{var} missing".
}
\examples{

my_survey <- data.frame(
  resp_id = 1:5,
  q1a = c(0, 1, 0, 1, 0),
  q1b = c(NA, NA, NA, 1, 0), # Asked if q1a \%in\% 1
  q2a = c(90, 80, 60, 40, 90),
  q2b = c("", "", NA, "Some reason for low rating", "") # Asked if q2a < 50
)

# Check that q1b has a value if and only if q1a \%in\% 1
\dontrun{
expect_base(q1b, q1a \%in\% 1, data = my_survey) # Fails for resp_id 2 and 5
}

# Check that q2b has a value if and only if q2a < 50
expect_base(q2b, q2a < 50, data = my_survey)

# Check that if q1a \%in\% 0 then q2a > 50 (but not vice-versa)
expect_cond(q1a \%in\% 0, q2a > 50, data = my_survey)

}
\seealso{
Other data expectations: 
\code{\link{datacomp-expectations}},
\code{\link{exclusivity-expectations}},
\code{\link{generic-expectations}},
\code{\link{uniqueness-expectations}},
\code{\link{value-expectations}}
}
\concept{data expectations}
