% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expect-data.R
\name{data-params}
\alias{data-params}
\title{Expectation params}
\arguments{
\item{var}{An unquoted column name to test.}

\item{vars}{<\code{\link[dplyr:dplyr_tidy_select]{tidy-select}}> A set of columns to
test.}

\item{flt}{<\code{\link[dplyr:dplyr_data_masking]{data-masking}}> A filter specifying
a subset of the data frame to test.}

\item{miss}{A vector of values to be treated as missing. The
\link[=testdat]{testdat.miss} option is used by default.}

\item{data}{A data frame to test. The \link[=global-data]{global test data} is used
by default.}
}
\value{
\verb{expect_*()} functions are mainly called for their side effects. The
expectation signals its result (e.g. "success", "failure"), which is logged
by the current \link[testthat:Reporter]{test reporter}. In a non-testing
context the expectation will raise an error with class
\code{expectation_failure} if it fails.
}
\description{
Expectation params
}
\keyword{internal}
