% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expect-datacomp.R
\name{datacomp-expectations}
\alias{datacomp-expectations}
\alias{expect_similar}
\alias{expect_valmatch}
\alias{expect_join}
\title{Expectations: cross-dataset expectations}
\usage{
expect_similar(var, data2, var2, flt = TRUE, flt2 = flt,
  threshold = 0.05, min = 100, data = get_testdata())

expect_valmatch(data2, vars, by, not = FALSE, flt = TRUE,
  data = get_testdata())

expect_join(data2, by = NULL, not = FALSE, flt = TRUE,
  data = get_testdata())
}
\arguments{
\item{var}{an unquoted variable name}

\item{data2}{the dataset to compare against}

\item{var2}{an unquoted variable name from data2}

\item{flt}{a filter specifying a subset of the dataset to test}

\item{flt2}{a filter specifying a subset of data2 to test}

\item{threshold}{the maximum proportional difference allowed between the two
categories}

\item{min}{the minimum number of responses for a category to allow
comparison. This avoidmall categories raising spurious errors}

\item{data}{data frame to test. The \link[=global-data]{global test data} is used
by default.}

\item{vars}{a list of columns generated by \code{\link[=vars]{vars()}}}

\item{by}{a character vector of variables to join by. For details see the man page for dplyr \link[=dplyr]{join}.}

\item{not}{reverse the results of the check}

\item{by}{a character vector of variables to join by. For details see the man page for dplyr \link[=dplyr]{join}.}
}
\description{
These functions allow for dataset comparisons
}
\seealso{
Other data expectations: \code{\link{conditional-expectations}},
  \code{\link{generic-expectations}},
  \code{\link{value-expectations}}
}
\concept{data expectations}
