% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expect-datacomp.R
\name{datacomp-expectations}
\alias{datacomp-expectations}
\alias{expect_similar}
\alias{expect_valmatch}
\alias{expect_join}
\title{Expectations: cross-dataset expectations}
\usage{
expect_similar(
  var,
  data2,
  var2,
  flt = TRUE,
  flt2 = flt,
  threshold = 0.05,
  min = 100,
  data = get_testdata()
)

expect_valmatch(
  data2,
  vars,
  by,
  not = FALSE,
  flt = TRUE,
  data = get_testdata()
)

expect_join(data2, by = NULL, not = FALSE, flt = TRUE, data = get_testdata())
}
\arguments{
\item{var}{an unquoted variable name}

\item{data2}{the dataset to compare against}

\item{var2}{an unquoted variable name from data2}

\item{flt}{a filter specifying a subset of the dataset to test}

\item{flt2}{a filter specifying a subset of data2 to test}

\item{threshold}{the maximum proportional difference allowed between the two
categories}

\item{min}{the minimum number of responses for a category to allow
comparison. This avoidmall categories raising spurious errors}

\item{data}{data frame to test. The \link[=global-data]{global test data} is used
by default.}

\item{vars}{a list of columns generated by \link[dplyr:vars]{vars()}}

\item{by}{a character vector of variables to join by. For details see dplyr
\link[dplyr:mutate-joins]{join()}.}

\item{not}{reverse the results of the check}
}
\description{
These functions allow for dataset comparisons:
\itemize{
\item \code{expect_similar} compares the distribution of a categorical variable
(\code{var}) from the one dataset (\code{data}) to that of a categorical variable
(\code{var2}) from another dataset (\code{data2}). The test fails if the
distributions are sufficiently dissimilar.
\item \code{expect_valmatch} compares the observations appearing in one dataset
(\code{data}) to the same observations, as picked out by a key (\code{by}), in
another dataset (\code{data2}). It fails if the selected variables (\code{vars})
aren't the same for those observations in both datasets.
\item \code{expect_join} compares one dataset (\code{data}) to another (\code{data2}) and
fails if all of the observations in the first, as picked out by a key
(\code{by}), do not appear in the second
}
}
\examples{

df1 <- data.frame(
  id = 0:99,
  binomial = sample(0:1, 100, TRUE),
  even = abs(0:99\%\%2 - 1) * 0:99
)

df2 <- data.frame(
  id = 0:99,
  binomial = sample(0:1, 100, TRUE),
  odd = 0:99\%\%2 *0:99
)


# Check categorical distribution is similar across data frames
\dontrun{expect_similar(binomial, df2, binomial, data = df1)}

# Check that same records 'succeeded' across data frames
\dontrun{expect_valmatch(df2, vars(binomial), by = "id", data = df1)}

# Check that a left join of df1 and df2 on "id" would not produce any NAs
expect_join(df2, by = "id", data = df1)

# Note:
expect_join(df2, by = c("even" = "id"), data = df1) # Passes
\dontrun{expect_join(df2, by = c("id" = "odd"), data = df1) # Fails}

}
\seealso{
Other data expectations: 
\code{\link{chk-expectations}},
\code{\link{conditional-expectations}},
\code{\link{exclusivity-expectations}},
\code{\link{generic-expectations}},
\code{\link{uniqueness-expectations}}
}
\concept{data expectations}
