% Generated by roxygen2 (4.0.0): do not edit by hand
\name{detect_outliers}
\alias{detect_outliers}
\title{Test for outliers}
\usage{
detect_outliers(dat, threshold = 1.96, plot = FALSE, ...)
}
\arguments{
\item{dat}{input dataset. Currently only supports \code{data.frame} but will soon support \code{data.table}}

\item{threshold}{for outlier detection method (defaults to z-score)}

\item{...}{Further args (to be passed to scores() function)}
}
\description{
This function identifies numeric columns in a data.frame, then identifies outliers and returns the indexes of these possible outliers for each numeric column.
}
\examples{
\dontrun{
data(iris)
detect_outliers(iris)
}
}

