% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expect-exclusive.R
\name{exclusivity-expectations}
\alias{exclusivity-expectations}
\alias{expect_exclusive}
\title{Expectations: exclusivity checks}
\usage{
expect_exclusive(vars, var_set, exc_val = 1, flt = TRUE, data = get_testdata())
}
\arguments{
\item{vars}{a list of columns generated by \link[dplyr:vars]{vars()}}

\item{var_set}{The full variable set to check, specified using \code{\link[=vars]{vars()}}. This
should include all variables specified in the \code{vars} argument}

\item{exc_val}{The value that flags a variable as "selected" (default: \code{1})}

\item{flt}{a filter specifying a subset of the dataset to test}

\item{data}{data frame to test. The \link[=global-data]{global test data} is used
by default.}
}
\description{
\code{expect_exclusive} tests that \code{vars} are exclusive - that, if any one of
\code{vars} is set to \code{exc_val}, no other variable in \code{vars} or \code{var_set} is also
set to \code{exc_val}.
}
\details{
This expectation is designed to check exclusivity in survey multiple response
sets, where one response is only valid on its own.

See the example data set below:
\itemize{
\item no record should have \code{q10_98}, "None of the above", selected while also
having any other response selected, so we refer to this an "exclusive"
response.
\item \code{expect_exclusive()} checks whether \code{q10_98} "None of the above" or
\code{q10_99} "Don't know", the exclusive responses, have been selected alongside
any other \verb{q10_*} response.
\item The expectations fails, since the first record has both \code{q10_1} and
\code{q10_98} selected.
}
}
\examples{

my_q_block <- data.frame(
  resp_id = 1:5, # Unique to respondent
  q10_1  = c(1, 1, 0, 0, 0),
  q10_2  = c(0, 1, 0, 0, 0),
  q10_3  = c(0, 0, 1, 0, 0),
  q10_98 = c(1, 0, 0, 1, 0), # None of the above
  q10_99 = c(0, 0, 0, 0, 1)  # Item not answered
)

# Make sure that if "None of the above" and "Item skipped" are selected
# none of the other question options are selected:
try(
expect_exclusive(
  c(q10_98, q10_99),
  starts_with("q10_"),
  data = my_q_block
)
)
}
\seealso{
Other data expectations: 
\code{\link{chk-expectations}},
\code{\link{conditional-expectations}},
\code{\link{datacomp-expectations}},
\code{\link{generic-expectations}},
\code{\link{uniqueness-expectations}}
}
\concept{data expectations}
