% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expect-deprec.R
\name{expect-deprec}
\alias{expect-deprec}
\alias{expect_func}
\alias{expect_join}
\title{Deprecated expectation functions}
\usage{
expect_func(var, ...)

expect_join(data2, by = NULL, not = FALSE, flt = TRUE, data = get_testdata())
}
\arguments{
\item{var}{an unquoted variable name}

\item{...}{arguments to pass to \code{expect_allany()}}

\item{data2}{the dataset to compare against}

\item{by}{a character vector of variables to join by. For details see dplyr
\link[dplyr:mutate-joins]{join()}.}

\item{not}{reverse the results of the check}

\item{flt}{a filter specifying a subset of the dataset to test}

\item{data}{data frame to test. The \link[=global-data]{global test data} is used
by default.}
}
\description{
\Sexpr[results=rd, stage=render]{testdat:::lifecycle("soft-deprecated")}

These functions are deprecated.

See \link[=generic-expectations]{Generic Expectation Functions} and \link[=datacomp-expectations]{Cross-dataset Expectations} for current expectation functions.
}
\seealso{
\link[=generic-expectations]{Generic Expectation Functions}

\link[=datacomp-expectations]{Cross-dataset Expectations}
}
\keyword{internal}
