% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprec-expect.R
\name{expect-deprec}
\alias{expect-deprec}
\alias{expect_func}
\alias{expect_join}
\alias{expect_similar}
\title{Deprecated expectation functions}
\usage{
expect_func(var, ...)

expect_join(data2, by = NULL, not = FALSE, flt = TRUE, data = get_testdata())

expect_similar(
  var,
  data2,
  var2,
  flt = TRUE,
  flt2 = flt,
  threshold = 0.05,
  min = 100,
  data = get_testdata()
)
}
\arguments{
\item{var}{an unquoted variable name}

\item{...}{arguments to pass to \code{expect_allany()}}

\item{data2}{the dataset to compare against}

\item{by}{a character vector of variables to join by. For details see dplyr
\link[dplyr:mutate-joins]{join()}.}

\item{not}{reverse the results of the check}

\item{flt}{a filter specifying a subset of the dataset to test}

\item{data}{data frame to test. The \link[=global-data]{global test data} is used
by default.}

\item{var2}{an unquoted variable name from data2}

\item{flt2}{a filter specifying a subset of data2 to test}

\item{threshold}{the maximum proportional difference allowed between the two
categories}

\item{min}{the minimum number of responses for a category to allow
comparison. This avoidmall categories raising spurious errors}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#soft-deprecated}{\figure{lifecycle-soft-deprecated.svg}{options: alt='[Soft-deprecated]'}}}{\strong{[Soft-deprecated]}}

These functions are deprecated.

See \link[=generic-expectations]{Generic Expectation Functions} and \link[=datacomp-expectations]{Cross-dataset Expectations} for current expectation functions.
}
\seealso{
\link[=generic-expectations]{Generic Expectation Functions}

\link[=datacomp-expectations]{Cross-dataset Expectations}
}
\keyword{internal}
