% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expect-deprec.R
\name{filter_expect}
\alias{filter_expect}
\title{Filter data to expectation result}
\usage{
filter_expect(data, expect_function, ..., not = TRUE)
}
\arguments{
\item{data}{A data frame to test.}

\item{expect_function}{An expectation function.}

\item{...}{Arguments to pass to expect_function.}

\item{not}{Reverse the results of the check.}
}
\value{
The input data frame filtered to records failing the expectation.
}
\description{
\Sexpr[results=rd, stage=render]{testdat:::lifecycle("defunct")}
}
\details{
This function is defunct
}
