% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporter-zzz.R
\name{global-data}
\alias{global-data}
\alias{set_testdata}
\alias{get_testdata}
\alias{context_data}
\title{Get/set test data}
\usage{
set_testdata(data)

get_testdata()

context_data(data)
}
\arguments{
\item{data}{data to be used}
}
\description{
A global test data set is used to avoid having to re-specify the testing
dataset in every test. These functions get and set the global data or set the
data for the current context.
}
\examples{
set_testdata(mtcars)
get_testdata()
}
