% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expect-labels.R
\name{label-expectations}
\alias{label-expectations}
\alias{chk_labels}
\alias{expect_labels}
\title{Testing data labels}
\usage{
chk_labels(x, val_labels = NULL, var_label = NULL)

expect_labels(
  var,
  val_labels = NULL,
  var_label = NULL,
  flt = TRUE,
  data = get_testdata()
)
}
\arguments{
\item{x}{vector to check}

\item{val_labels}{a character vector of expected labels or a named vector of
expected label-value pairs.}

\item{var_label}{a character vector, the variable label to test for}

\item{var}{an unquoted variable name}

\item{flt}{a filter specifying a subset of the dataset to test}

\item{data}{data frame to test. The \link[=global-data]{global test data} is used
by default.}
}
\value{
A logical vector flagging records that have passed or failed the check
}
\description{
Testing data labels
}
\examples{

\dontrun{
  df <- data.frame(
    x = labelled::labelled(c("M", "M", "F"), c(Male = "M", Female = "F"), "Sex"),
    y = labelled::labelled(c("M", "M", "F"), c(Male = "M", Female = "F", Other = "X"))
  )

  expect_labels(x, c(Male = "M"), data = df)
  expect_labels(x, labelled::val_labels(df$y), data = df) # N.B. This passes!
  expect_labels(x, "Male", data = df)
  expect_labels(x, var_label = "Sex", data = df)
}

}
