% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprec-reporter.R
\name{reporter-deprec}
\alias{reporter-deprec}
\alias{context_data}
\title{Deprecated reporter functions}
\usage{
context_data(data)
}
\arguments{
\item{data}{Data frame to be used.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#soft-deprecated}{\figure{lifecycle-soft-deprecated.svg}{options: alt='[Soft-deprecated]'}}}{\strong{[Soft-deprecated]}}

These functions are deprecated.

See \link[=global-data]{Get/set test data} for working with global test data.
}
\seealso{
\link[=global-data]{Get/set test data}
}
\keyword{internal}
