% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testdat-package.R
\docType{package}
\name{testdat-package}
\alias{testdat}
\alias{testdat-package}
\title{testdat: Data Unit Testing for R}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Test your data! An extension of the 'testthat' unit testing framework with a family of functions and reporting tools for checking and validating data frames.
}
\section{Options}{

\itemize{
\item \code{testdat.miss}: A vector of values to consider missing
(default: \code{c(NA, "")}).
\item \code{testdat.miss_text}: A vector of values to consider missing in text
variables
(default: \code{c("error", "null", "0", ".", "-", ",", "na", "#n/a", "", NA)}).
\item \code{testdat.stop_on_fail}: Should an expectation raise an error on failure?
Useful for interactive use of expectation functions (default: \code{TRUE}).
\item \code{testdat.scipen}: When it is necessary to convert a numeric vector to
character for checking, this value will be used for \code{scipen}
(default: \code{999}).
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://socialresearchcentre.github.io/testdat/}
  \item \url{https://github.com/socialresearchcentre/testdat}
  \item Report bugs at \url{https://github.com/socialresearchcentre/testdat/issues}
}

}
\author{
\strong{Maintainer}: Danny Smith \email{danny@gorcha.org}

Authors:
\itemize{
  \item Kinto Behr \email{kinto.behr@srcentre.com.au}
}

Other contributors:
\itemize{
  \item The Social Research Centre [copyright holder]
}

}
\keyword{internal}
