% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expect-value.R
\name{value-expectations}
\alias{value-expectations}
\alias{expect_values}
\alias{expect_regex}
\alias{expect_range}
\alias{expect_exclusive}
\alias{expect_unique}
\alias{expect_unique_across}
\title{Expectations: value checks}
\usage{
expect_values(var, ..., miss = getOption("testdat.miss"),
  data = get_testdata())

expect_regex(var, pattern, flt = TRUE, data = get_testdata())

expect_range(var, min, max, flt = TRUE, data = get_testdata())

expect_exclusive(vars, exc_vars, exc_val = 1, flt = TRUE,
  data = get_testdata())

expect_unique(vars, flt = TRUE, data = get_testdata())

expect_unique_across(vars, flt = TRUE, data = get_testdata())
}
\arguments{
\item{var}{an unquoted variable name}

\item{...}{vectors of valid values}

\item{miss}{a vector of values to be treated as missing. The
\link[=testdat]{testdat.miss} option is used by default}

\item{data}{data frame to test. The \link[=global-data]{global test data} is used
by default.}

\item{pattern}{a regex to check}

\item{flt}{a filter specifying a subset of the dataset to test}

\item{min}{minimum value for range check}

\item{max}{maximum value for range check}

\item{vars}{a list of columns generated by \code{\link[=vars]{vars()}}}

\item{exc_vars}{The variable set to check, specified using \code{\link[=vars]{vars()}}. This
should include all variables specified in the \code{vars} argument}

\item{exc_val}{The value to check for exclusivity (default: \code{1})}
}
\description{
These functions test variable values
}
\seealso{
Other data expectations: \code{\link{conditional-expectations}},
  \code{\link{datacomp-expectations}},
  \code{\link{generic-expectations}}
}
